src/main/java/com/mx/manuel/mockservice/config/HandleError.java
package com.mx.manuel.mockservice.config;

import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;
import org.springframework.web.bind.annotation.ExceptionHandler;

@ControllerAdvice
public class HandleError {

    private static final Logger LOG = LoggerFactory.getLogger(HandleError.class);

    @ExceptionHandler(Exception.class)
    public ResponseEntity<?> handleUnhandledError(Exception ex) {
        LOG.error("Error: " + ex);
        return new ResponseEntity<>(ex.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
    }

    @ExceptionHandler(IllegalArgumentException.class)
    public ResponseEntity<?> handleEntityNotFound(IllegalArgumentException ex) {
        return new ResponseEntity<>(ex.getMessage(), HttpStatus.NOT_FOUND);
    }
}
